+++ fcp/src/cmd_read.c
@@ -80,12 +80,12 @@ static int __read(args_t * args, off_t offset, entry_list_t * done_list)
 	if (ffs->count <= 0)
 		return 0;
 
-	size_t block_size;
+	uint32_t block_size;
 	if (__ffs_info(ffs, FFS_INFO_BLOCK_SIZE, &block_size) < 0)
 		return -1;
 
 	if (args->buffer != NULL) {
-		size_t buffer;
+		uint32_t buffer;
 		if (parse_size(args->buffer, &buffer) < 0)
 			return -1;
 		if (__ffs_buffer(ffs, buffer) < 0)
@@ -101,7 +101,7 @@ static int __read(args_t * args, off_t offset, entry_list_t * done_list)
 	if (entry.actual == 0) {
 		if (args->verbose == f_VERBOSE)
 			fprintf(stderr, "%8llx: %s: is empty (skip)\n",
-		       		offset, name);
+		       		(long long)offset, name);
 		return 0;
 	}
 
@@ -113,14 +113,14 @@ static int __read(args_t * args, off_t offset, entry_list_t * done_list)
 	if (entry.type == FFS_TYPE_LOGICAL) {
 		if (args->verbose == f_VERBOSE)
 			fprintf(stderr, "%8llx: %s: logical (skip)\n",
-		       		offset, name);
+		       		(long long)offset, name);
 		return 0;
 	}
 
 	if (entry_list_exists(done_list, &entry) == 1) {
 		if (args->verbose == f_VERBOSE)
 			fprintf(stderr, "%8llx: %s: write to '%s' (skip)\n",
-			        offset, full_name, out_path);
+			        (long long)offset, full_name, out_path);
 		return 0;
 	}
 
@@ -148,7 +148,7 @@ static int __read(args_t * args, off_t offset, entry_list_t * done_list)
 
 		if (args->verbose == f_VERBOSE)
 			fprintf(stderr, "%8llx: %s: write to '%s' (done)\n",
-		       		offset, full_name, out_path);
+		       		(long long)offset, full_name, out_path);
 	}
 
 	return 0;
